+++ ElvUI_EltreumUI/Modules/DarkLightColors.lua
@@ -23,10 +23,10 @@ function ElvUI_EltreumUI:DarkMode()
 		E.db["auras"]["buffs"]["barTexture"] = "Eltreum-Blank"
 		E.db["auras"]["debuffs"]["barTexture"] = "Eltreum-Blank"
 		E.db["unitframe"]["statusbar"] = "Eltreum-Blank"
-		E.db["unitframe"]["colors"]["transparentAurabars"] = false
-		E.db["unitframe"]["colors"]["transparentCastbar"] = false
-		E.db["unitframe"]["colors"]["transparentHealth"] = false
-		E.db["unitframe"]["colors"]["transparentPower"] = false
+		E.db["unitframe"]["colors"]["transparentAurabars"] = true
+		E.db["unitframe"]["colors"]["transparentCastbar"] = true
+		E.db["unitframe"]["colors"]["transparentHealth"] = true
+		E.db["unitframe"]["colors"]["transparentPower"] = true
 
 		--setup namecolors
 		if ElvUI_EltreumUI.Classic or ElvUI_EltreumUI.TBC then
@@ -79,9 +79,9 @@ function ElvUI_EltreumUI:LightMode()
 		E.db["unitframe"]["units"]["assist"]["colorOverride"] = "USE_DEFAULT"
 		E.db["unitframe"]["units"]["target"]["colorOverride"] = "USE_DEFAULT"
 		E.db["unitframe"]["units"]["targettarget"]["colorOverride"] = "USE_DEFAULT"
-		E.db["auras"]["buffs"]["barTexture"] = (unitframeclass[E.myclass])
-		E.db["auras"]["debuffs"]["barTexture"] = (unitframeclass[E.myclass])
-		E.db["unitframe"]["statusbar"] = (unitframeclass[E.myclass])
+		E.db["auras"]["buffs"]["barTexture"] = "Eltreum-Blank"
+		E.db["auras"]["debuffs"]["barTexture"] = "Eltreum-Blank"
+		E.db["unitframe"]["statusbar"] = "Eltreum-Blank"
 		E.db["unitframe"]["colors"]["transparentAurabars"] = false
 		E.db["unitframe"]["colors"]["transparentCastbar"] = false
 		E.db["unitframe"]["colors"]["transparentHealth"] = false